feat: allow FE import from CSV
