+++ prisma/seed.ts
@@ -1,5 +1,5 @@
 import { signPassword } from '@/services/auth'
-import getEmissionFactors from '@/services/baseEmpreinte/getEmissionFactors'
+import { getEmissionFactorsFromAPI } from '@/services/baseEmpreinte/getEmissionFactorsFromAPI'
 import { reCreateBegesRules } from '@/services/exportRules/beges'
 import { faker } from '@faker-js/faker'
 import { EmissionFactorStatus, Import, Level, PrismaClient, Role, StudyRole, SubPost, Unit, User } from '@prisma/client'
@@ -249,7 +249,7 @@ const licenses = async () => {
 const main = async (params: Params) => {
   await Promise.all([actualities(), licenses(), users(), reCreateBegesRules()])
   if (params.importFactors) {
-    await getEmissionFactors({ name: params.importFactors })
+    await getEmissionFactorsFromAPI(params.importFactors)
   }
 }
