+++ src/scripts/baseEmpreinte/getEmissionFactors.ts
@@ -1,5 +1,6 @@
 import { Command } from 'commander'
-import getEmissionFactors from '../../services/baseEmpreinte/getEmissionFactors'
+import { getEmissionFactorsFromAPI } from '../../services/baseEmpreinte/getEmissionFactorsFromAPI'
+import { getEmissionFactorsFromCSV } from '../../services/baseEmpreinte/getEmissionFactorsFromCSV'
 
 const program = new Command()
 
@@ -8,6 +9,13 @@ program
   .description("Script pour importer les facteurs d'émission depuis la base empreinte")
   .version('1.0.0')
   .requiredOption('-n, --name <value>', 'Nom de la version')
+  .option('-f, --file <value>', 'Import from CSV file')
   .parse(process.argv)
 
-getEmissionFactors(program.opts())
+const params = program.opts()
+
+if (params.file) {
+  getEmissionFactorsFromCSV(params.name, params.file)
+} else {
+  getEmissionFactorsFromAPI(params.name)
+}
